+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/export/processor.py
@@ -14,7 +14,7 @@
 
 from typing import Sequence
 
-from opentelemetry.sdk.metrics.export import MetricRecord
+from opentelemetry.sdk.metrics.export import ExportRecord
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.util import get_dict_as_key
 
@@ -36,22 +36,22 @@ class Processor:
         self.stateful = stateful
         self._resource = resource
 
-    def checkpoint_set(self) -> Sequence[MetricRecord]:
-        """Returns a list of MetricRecords used for exporting.
+    def checkpoint_set(self) -> Sequence[ExportRecord]:
+        """Returns a list of ExportRecords used for exporting.
 
-        The list of MetricRecords is a snapshot created from the current
+        The list of ExportRecords is a snapshot created from the current
         data in all of the aggregators in this processor.
         """
-        metric_records = []
+        export_records = []
         # pylint: disable=W0612
         for (
             (instrument, aggregator_type, _, labels),
             aggregator,
         ) in self._batch_map.items():
-            metric_records.append(
-                MetricRecord(instrument, labels, aggregator, self._resource)
+            export_records.append(
+                ExportRecord(instrument, labels, aggregator, self._resource)
             )
-        return metric_records
+        return export_records
 
     def finished_collection(self):
         """Performs certain post-export logic.