+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/export/aggregate.py
@@ -34,6 +34,7 @@ class Aggregator(abc.ABC):
         self._lock = threading.Lock()
         self.last_update_timestamp = 0
         self.initial_checkpoint_timestamp = 0
+        self.first_timestamp = time_ns()
         self.checkpointed = True
         if config is not None:
             self.config = config